Setup a public API for the echosms module
