+++ src/echosms/__init__.py
@@ -1,0 +1,7 @@
+"""Setup the public API for echoSMs."""
+from .benchmarkdata import BenchMarkData
+from .referencemodels import ReferenceModels
+from .scattermodels import MSSModel, PSMSModel
+from .utils import Utils
+
+__all__ = ['BenchMarkData', 'ReferenceModels', 'MSSModel', 'PSMSModel', 'Utils']
